test(common): split perf counter JSON writer test
